chaps: Limit GenerateRandom() size to prevent out of memory
BUG=b:299372668
TEST=unittest

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4889284
Commit-Queue: John L Chen <zuan@chromium.org>
Tested-by: John L Chen <zuan@chromium.org>

+++ chaps/session.h
@@ -119,7 +119,7 @@ class Session {
                                 int* new_private_key_handle) = 0;
   // Random number generation (see PKCS #11 v2.20: 11.15).
   virtual CK_RV SeedRandom(const std::string& seed) = 0;
-  virtual CK_RV GenerateRandom(int num_bytes, std::string* random_data) = 0;
+  virtual CK_RV GenerateRandom(size_t num_bytes, std::string* random_data) = 0;
   // Returns true if private objects are loaded and the session is ready for
   // operations with them without blocking.
   virtual bool IsPrivateLoaded() = 0;
